Another pack of improvements and fixes (#10)
* Another pack of improvements and fixes

-Added Fireworks and Firework Box to Red Envelope drops cause it makes sense
-Added Red Envelope as an extra drop to Mythical Wyvern

* Updated Mod descriptions (DEVS REVIEW REQUIRED)

Updated Mod descriptions with nicknames of new contributors and rewrote the whole text to be more clear & user-friendly. Some changes might be controversial so I would suggest the devs to check if my suggested design is ok before merging

* Added 3 Vanity Sets from Heartbeataria mod to add them in Red Envelope drop pool in the future

Added 3 Vanity Sets from Heartbeataria mod to add them in Red Envelope drop pool in the future as they better fit there as a special drop instead of how they work in Heartbeataria

* -Changed Red Envelope drop system

Changed Red Envelope drop system to make it more similar to vanilla Goodie Bags, provide more variety in it's drops and include the 3 new vanity sets.

New loottable looks like this:
-Fiery Cracker (10 to 20): 10%
-Release Lantern (3 to 10): 10%
-Red Rocket (3 to 15): 10% (new drop)
-Green Rocket (3 to 15): 10% (new drop)
-Blue Rocket (3 to 15): 10% (new drop)
-Yellow Rocket (3 to 15): 10% (new drop)
-Firework Box: 8% (new drop)
-Mythical vanity set: 8%
-Tang Suit vanity set: 8% (new drop)
-Mythical Dog vanity set: 8% (new drop)
-Tang Yuan vanity set: 8% (new drop)
-Silver coins (5 to 15): 100%

Each of the possible drops can spawn once from each Red Envelope, besides the coins that spawn paired with other loot. The code that I wrote to make this system is probably terribly bad and inefficient from programming perspective, but it works flawlessly in-game.

* Updated Rainbow Piece tooltip

Updated Rainbow Piece tooltip to be more in-pair with rest of the vanilla ones

* Fixed "Soul of Blight" and a bunch of Bestiary entries

- Fixed "Soul of Blight" sprite and it's items to be more in-pair with vanilla ones
- Fixed a bunch of typos and misspelled words in different Bestiary entries

* Reverted some of the mistakely-done changes